+++ src/main/java/com/raoulvdberge/refinedstorage/api/network/security/ISecurityCardContainer.java
@@ -1,5 +1,6 @@
 package com.raoulvdberge.refinedstorage.api.network.security;
 
+import javax.annotation.Nullable;
 import java.util.List;
 
 /**
@@ -7,7 +8,13 @@ import java.util.List;
  */
 public interface ISecurityCardContainer {
     /**
-     * @return the security cards in this container
+     * @return the security cards in this container, {@link ISecurityCard#getOwner()} CANNOT be null for any card in this list!
      */
     List<ISecurityCard> getCards();
+
+    /**
+     * @return the global security card in this container, or null if there is none, {@link ISecurityCard#getOwner()}) can be null!
+     */
+    @Nullable
+    ISecurityCard getGlobalCard();
 }